fix(search): restore input focus when Vaul steals it during snap transitions
When Vaul/Radix focuses the drawer container (e.g. during a mobile keyboard-
triggered viewport resize causing a snap transition), the previous focusin-
only blur sent focus to BODY, causing the input's onBlur to fire and close the
recents dropdown after 150ms.

Fix: move the focus guard to Map.tsx (where mapSearchInputRef is accessible).
The focusin capture listener now checks e.relatedTarget — if the input was the
element that just lost focus to the drawer, restore it immediately. Also cancel
any pending blurTimeoutRef in the input's onFocus so a brief focus-bounce
doesn't trigger the 150ms close timer.
